services/data-storage/requirements.txt
paho-mqtt
SQLAlchemy
fastapi
uvicorn